fintech-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pillow
pytesseract
python-dotenv
